ensured that division return as int: data  type casting

+++ 0x02-python-import_modules/1-calculation.py
@@ -8,4 +8,4 @@ b = 5
 print("{} + {} = {}".format(a, b, a + b))
 print("{} - {} = {}".format(a, b, a - b))
 print("{} * {} = {}".format(a, b, a * b))
-print("{} / {} = {}".format(a, b, a / b))
+print("{} / {} = {}".format(a, b, int(a / b)))
